Change ClientName and improve file download condition
ClientName has been changed from "GamerTech" to "AztexCraft" in the unit tests and conditions for downloading files in the GmlClientManager have been improved. Now, a file will only be skipped during the download process if it exists in both the whitelist files and the local path. This change ensures all essential files are present for each profile.