Update path handling and monitoring in file watcher
The main focus of this commit is enhancing how directories are managed and how file changes are detected in the Gml.Client project. The SystemIOProcedures class now specifies a list of paths to be monitored, instead of just including the profile directory. Furthermore, in ApiProcedures, the handling of ProfileFileWatcher instances has been revised to differentiate between mod and asset directories. Changes in the ProfileFileWatcher class include adding a flag to determine whether changes should cause the process to be killed.