Dev (#14)
* Add hardware ID parameter in authentication method

The Auth method was updated to include a new parameter, the hardware ID (hwid), to strengthen the security of the login process in the Gml.Client. Consequently, the Auth method signature was also amended across the ApiProcedures, IGmlClientManager and GmlClientManager classes to accommodate this change.

* Updated UnitTest setup to use environment variables

The UnitTest setup included hardcoded paths for localDirectory and baseAddress. These have now been updated to use the Environment.CurrentDirectory for the localDirectory and a new base https address for the baseAddress, improving flexibility and portability of the tests.

* Update parameters order in GmlClientManager instantiation

The parameters for the creation of the GmlClientManager in UnitTests.cs has been reordered. The localDirectory parameter now comes before the baseAddress parameter. This update is necessary for proper initialization and functioning of the GmlClientManager.

* Add OS type information to profile and refactor path adjustments (#11)

Additional operating system (OS) type has been added to profile information. Path adjustments related to file download and directory specification have been refactored. For the downloaded files, the listed order has been updated to descending based on the file size. Additional precautions are in place for process start depending on the OS type.

Co-authored-by: GamerVII-NET <gamervii.phone@gmail.com>

* Add OS type information to profile and refactor path adjustments (#13)

Additional operating system (OS) type has been added to profile information. Path adjustments related to file download and directory specification have been refactored. For the downloaded files, the listed order has been updated to descending based on the file size. Additional precautions are in place for process start depending on the OS type.

Co-authored-by: GamerVII-NET <gamervii.phone@gmail.com>

---------

Co-authored-by: GamerVII-NET <gamervii.phone@gmail.com>