Merge dev and master branches (#17)
* Add hardware ID parameter in authentication method

The Auth method was updated to include a new parameter, the hardware ID (hwid), to strengthen the security of the login process in the Gml.Client. Consequently, the Auth method signature was also amended across the ApiProcedures, IGmlClientManager and GmlClientManager classes to accommodate this change.

* Updated UnitTest setup to use environment variables

The UnitTest setup included hardcoded paths for localDirectory and baseAddress. These have now been updated to use the Environment.CurrentDirectory for the localDirectory and a new base https address for the baseAddress, improving flexibility and portability of the tests.

* Update parameters order in GmlClientManager instantiation

The parameters for the creation of the GmlClientManager in UnitTests.cs has been reordered. The localDirectory parameter now comes before the baseAddress parameter. This update is necessary for proper initialization and functioning of the GmlClientManager.

* Add OS type information to profile and refactor path adjustments (#11)

Additional operating system (OS) type has been added to profile information. Path adjustments related to file download and directory specification have been refactored. For the downloaded files, the listed order has been updated to descending based on the file size. Additional precautions are in place for process start depending on the OS type.

Co-authored-by: GamerVII-NET <gamervii.phone@gmail.com>

* Add OS type information to profile and refactor path adjustments (#13)

Additional operating system (OS) type has been added to profile information. Path adjustments related to file download and directory specification have been refactored. For the downloaded files, the listed order has been updated to descending based on the file size. Additional precautions are in place for process start depending on the OS type.

Co-authored-by: GamerVII-NET <gamervii.phone@gmail.com>

* Dev with splashscreen (#16)

* Remove Gml.Client Interfaces and Enums

All interfaces and enums under src/Gml.Client/src/GmlCore.Interfaces were removed. This removal includes all enums, launcher, procedures, storage, and user interfaces. The changes also involve some adjustments in IntegrationUpdateDto.cs, ProfileCreateDto.cs, ProfileUpdateDto.cs and the introduction of the StreamExtensions class.

* Add new integrations, authentication, and procedures

This commit introduces a variety of enhancements including new integrations (like AuthCustomResponse, AuthResult), enums (like AuthType and GameLoader), and procedures (like IFileStorageProcedures and IUserProcedures). Moreover, it adds classes and interfaces for various system actions, launcher functions, and storage settings.

* Replace progress event with Reactive Extensions

The EventHandler<ProgressChangedEventArgs> has been replaced with an ISubject<int> from Reactive Extensions in API procedures. This allows for using an IObservable in the GmlClientManager. In addition, the error file sorting order has been changed in SystemIOProcedures.

---------

Co-authored-by: GamerVII-NET <gamervii.phone@gmail.com>