Merge dev and master branches (#24)
* Increase concurrent download limit and correct path parameters

The concurrent download limit in the DownloadFiles method was increased from 16 to 60 for efficiency. Also, the incorrect 'profilePath' parameter was replaced with 'installationDirectory' in ApiProcedures class, to ensure correct paths are used in the process start info and parameters dictionary.

* Remove unnecessary newline in ApiProcedures

This commit removes a superfluous newline in the ApiProcedures file. This brings the code into closer alignment with our established code style guidelines and improves readability.